requirements.txt
apache-airflow
apache-airflow-providers-amazon
apache-airflow-providers-snowflake
astro-sdk-python[amazon,snowflake]
astronomer-cosmos[dbt.snowflake]
loguru
pandas
pre-commit
python-dotenv
requests
snowflake-connector-python
sqlfluff
